% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf-palette.R
\name{mf_pal}
\alias{mf_pal}
\title{mf color palette}
\usage{
mf_pal(reverse = FALSE, ...)
}
\arguments{
\item{fill}{Use the fill palette.}
}
\description{
Creates a palette function that accepts an integer argument (the number of
levels in the scale) and returns a vector of corresponding hex values.
}
\details{
mf_pal is a utility function for scale_color_
}
\seealso{
Other colour mf: 
\code{\link{scale_colour_mf}()}
}
\concept{colour mf}
