% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf-palette.R
\name{theme_mf}
\alias{theme_mf}
\alias{theme_mf_white}
\title{ggplot color theme based on MF's preferences}
\usage{
theme_mf(
  base_size = 10,
  base_family = "sans",
  horizontal = TRUE,
  dkpanel = FALSE
)

theme_mf_white(
  base_size = 11,
  base_family = "sans",
  gray_bg = TRUE,
  horizontal = TRUE
)
}
\arguments{
\item{horizontal}{\code{logical} Horizontal axis lines?}

\item{dkpanel}{\code{logical} Darker background for panel region?}

\item{gray_bg}{\code{logical} If \code{TRUE}, use gray background, else
   use white
background.}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A theme in the style preferences of \emph{MF}.
}
\details{
\code{theme_mf} implements the standard bluish-gray
background theme in the print \emph{The Economist} and
\href{http://economist.com}{economist.com}.

\code{theme_economist_white} implements a variant with a while
panel and light gray (or white) background often used by \emph{The Economist}
blog \href{http://www.economist.com/blogs/graphicdetail}{Graphic Detail}.

Use \code{\link{scale_color_mf}()} with this theme.
The x axis should be displayed on the right hand side.

\emph{The Economist} uses "ITC Officina Sans" as its font for graphs. If
you have access to this font, you can use it with the
\pkg{extrafont} package. "Verdana" is a good substitute.
}
\references{
\itemize{
\item \href{http://economist.com}{The Economist}
\item \href{http://spiekermann.com/en/itc-officina-display/}{Spiekerblog, "ITC Officina Display", January 1, 2007.}
\item \url{http://www.economist.com/help/about-us}
}
}
\concept{themes economist}
